Use `Puppet::Util::Execution.execute`
When the puppetserver uses jruby 9k, this report processor often fails
with...

```
Puppet Report processor failed: Cannot allocate memory - hostname
```

This is because the entire jvm process is forked just in order to run
the `hostname` command.  JRuby9k is the default in Puppet Server 6 and
has also been the default in Puppet Enterprise for a while now.

The issue is especially noticable if Puppet Server is configured with a
large Java heap. (eg. when running with an 8GB heap, a whole extra 8GB
will be allocated when forking the process).